requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.61.1
django-filter==23.5
django-phonenumber-field==7.3.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.6
Markdown==3.5.2
mysqlclient==2.2.4
oauthlib==3.2.2
phonenumbers==8.13.31
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
